Able to delete JEV

+++ ACC.Data/RolesRepository.cs
@@ -131,7 +131,7 @@ namespace ACC.Data
                     {
                         var parameters = new object[][]
                         {
-                            new object[] { "@id", DbType.Int16, entity.Id},
+                            new object[] { "@id", DbType.Int32, entity.Id},
                         };
 
                         string query = $"DELETE FROM {tableName} WHERE id = @id";
